Rename oauth2-client to app-client
OAuth2 client is way to confusing.
