+++ src/routes.ts
@@ -6,10 +6,10 @@ import apps from './app/controller/collection.js';
 import blob from './blob/controller.js';
 import changePassword from './changepassword/controller.js';
 import changePasswordRedirect from './well-known/controller/change-password.js';
-import client from './oauth2-client/controller/item.js';
-import clientEdit from './oauth2-client/controller/edit.js';
-import clientNew from './oauth2-client/controller/new.js';
-import clients from './oauth2-client/controller/collection.js';
+import client from './app-client/controller/item.js';
+import clientEdit from './app-client/controller/edit.js';
+import clientNew from './app-client/controller/new.js';
+import clients from './app-client/controller/collection.js';
 import group from './group/controller/item.js';
 import groupNew from './group/controller/new.js';
 import groups from './group/controller/collection.js';
