+++ src/services.ts
@@ -4,3 +4,4 @@ export * as user from './user/service.js';
 export * as privilege from './privilege/service.js';
 export * as resetPassword from './reset-password/service.js';
 export * as log from './log/service.js';
+export * as appClient from './app-client/service.js';
